+++ components/ChartData.tsx
@@ -8,6 +8,10 @@ import gql from "graphql-tag";
 import { useQuery } from "@apollo/react-hooks";
 import TablePagination from "@material-ui/core/TablePagination";
 import numbro from "numbro";
+import ComponentLoader from './misc/ComponentLoader';
+import NotAvailable from './misc/NotAvailable'
+import ErrorMessage from './misc/ErrorMessage';
+
 import { GET_CHAIN } from './query/Chain'
 
 
@@ -42,10 +46,6 @@ const useStyles = makeStyles({
   },
 });
 
-const NotAvailable = () => {
-  const classes = useStyles();
-  return <span className={classes.label}>Data currently not available</span>
-}
 
 const ChartData = () => {
   const classes = useStyles();
@@ -54,8 +54,8 @@ const ChartData = () => {
     pollInterval: 5000,
   });
 
-  if (loading) return <>{"Loading..."}</>
-  if (error) return <>{`Error! ${error.message}`}</>
+  if (loading) return <ComponentLoader size="small" />
+  if (error) return <ErrorMessage message={error.message} />
 
 
   return (
@@ -70,7 +70,7 @@ const ChartData = () => {
               <Typography variant="h5" className={classes.value}>
                 $ {numbro(data.chain.tokenPrice.usd).format("0.00")}
               </Typography>
-              : <NotAvailable />}
+              : <NotAvailable variant="body2" className={classes.value} />}
           </Card>
         </Grid>
 
@@ -83,7 +83,7 @@ const ChartData = () => {
               <Typography variant="h5" className={classes.value}>
                 $ {numbro(data.chain.tokenPrice.usdMarketCap).format("0.00")}
               </Typography>
-              : <NotAvailable />}
+              : <NotAvailable variant="body2" className={classes.value} />}
           </Card>
         </Grid>
 
@@ -96,7 +96,7 @@ const ChartData = () => {
               <><Typography variant="h5" className={classes.value}>
                 {numbro(data.chain.averageBlockTime).format("0.00")}
               </Typography><span className={classes.valueSuffix}>seconds</span></>
-              : <NotAvailable />}
+              : <NotAvailable variant="body2" className={classes.value} />}
           </Card>
         </Grid>
 
@@ -109,7 +109,7 @@ const ChartData = () => {
               <Typography variant="h5" className={classes.value}>
                 {numbro(data.chain.txCount).format("000,000")}
               </Typography>
-              : <NotAvailable />}
+              : <NotAvailable variant="body2" className={classes.value} />}
           </Card>
         </Grid>
 
@@ -122,7 +122,7 @@ const ChartData = () => {
               <Typography variant="h5" className={classes.value}>
                 {numbro(data.chain.latestHeight).format("000,000")}
               </Typography>
-              : <NotAvailable />}
+              : <NotAvailable variant="body2" className={classes.value} />}
           </Card>
         </Grid>
 
@@ -135,7 +135,7 @@ const ChartData = () => {
               <Typography variant="h5" className={classes.value}>
                 {numbro(data.chain.walletCount).format("000,000")}
               </Typography>
-              : <NotAvailable />}
+              : <NotAvailable variant="body2" className={classes.value} />}
           </Card>
         </Grid>
       </Grid>