+++ components/ledger/validatorGroup/vote/Confirm.tsx
@@ -8,6 +8,7 @@ import Typography from "@material-ui/core/Typography";
 import Grid from "@material-ui/core/Grid";
 import Divider from "@material-ui/core/Divider";
 import IconButton from "@material-ui/core/IconButton";
+import NotAvailable from '../../../misc/NotAvailable'
 
 
 const useStyles = makeStyles({
@@ -125,7 +126,7 @@ const Confirm = () => {
                                     Michelle Clark
                 {/* {data.block && data.block.timestamp
                 ? new Date(parseInt(data.block.timestamp) * 1000).toUTCString()
-                : "Data currently not available"}{" "}
+                : <NotAvailable variant="body2" />}
               (
               {data && data.block && data.block.timestamp
                 ? moment.unix(data.block.timestamp).fromNow()
@@ -146,7 +147,7 @@ const Confirm = () => {
               data.block.transactions &&
               data.block.transactions.transactionIndex
                 ? data.block.transactions.transactionIndex.length()
-                : "Data currently not available"} */}
+                : <NotAvailable variant="body2" />} */}
                                 </Typography>
                                 <Divider variant="middle" className={classes.divider} />
                             </Grid>
@@ -159,7 +160,7 @@ const Confirm = () => {
                                     1 Locked CGLD
                 {/* {data.block && data.block.size
                 ? data.block.size
-                : "Data currently not available"} */}
+                : <NotAvailable variant="body2" />} */}
                                 </Typography>
                                 <Divider variant="middle" className={classes.divider} />
                             </Grid>