+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 * Implemented Ledger Dialog Screens 
 * [#5] Fixed Functional components by adding JSX Tags to return value
 * [#3] Updated attribute element to CamelCase
-* [#25] Grouped all gql queries into single directory 'query' 
+* [#23] Added NotAvailable, ErrorMessage, ComponentLoader components in /misc
+* [#24] Implemented ComponentLoader in /misc
+* [#25] Grouped all gql queries into single directory 'query'
 * [#28] Fixed the `build-fragement` npm script by runing it as Javscript
 * [#35] Fixed error 'block number is not assignable to type 'string' ' by parsing its value to a string